Replace Yandex search test with Bing search test
